+++ db/migrations/20210320000001.up.sql
@@ -1,0 +1,1 @@
+alter table project_index add column num_tries integer default 0;