+++ src/logging.lisp
@@ -15,7 +15,9 @@
 (defparameter *max-depth* 100)
 
 
-(defun setup (log-dir &key (level :error))
+(defun setup (log-dir &key
+                        (level :error)
+                        (app "app"))
   (setf *log-dir* log-dir)
 
   (setf log4cl-extras/error:*max-traceback-depth*
@@ -24,18 +26,21 @@
   (log4cl-extras/config:setup
     `(:level ,level
       :appenders ((daily :layout :json
-                         :name-format ,(format nil "~A/app.log"
-                                               log-dir)
+                         :name-format ,(format nil "~A/~A.log"
+                                               log-dir
+                                               app)
                          :backup-name-format "app-%Y%m%d.log")))))
 
 
-(defun setup-for-repl (&key (level :debug) (app "app"))
+(defun setup-for-repl (&key
+                         (level :debug)
+                         (app "app"))
   (setf log4cl-extras/error:*max-traceback-depth*
         *max-depth*)
   
   (let ((filename (format nil "~A/~A.log"
                           (or *log-dir*
-                              "/tmp")
+                              "/app/logs")
                           app))
         (backup-filename (format nil "~A-%Y%m%d.log"
                                  app)))